Finalizing application
Finshed adding functionality to all of the buttons.
Made the display screen properly show the previous operand with the chosen
operator beside it.
When an operation is ready and the user presses another operator after hitting a
number, the operands will be evauated before adding the new operator to the
display.
Calculator is complete. Final commit.